Split vector tile classes into separate files
