+++ src/core/vectortile/qgsvectortilelayer.cpp
@@ -22,10 +22,8 @@
 #include "qgsvectortilebasiclabeling.h"
 #include "qgsvectortilebasicrenderer.h"
 #include "qgsvectortilelabeling.h"
-#include "qgsvectortileloader.h"
 #include "qgsvectortileutils.h"
 #include "qgsnetworkaccessmanager.h"
-
 #include "qgsdatasourceuri.h"
 #include "qgslayermetadataformatter.h"
 #include "qgsblockingnetworkrequest.h"
@@ -39,10 +37,11 @@
 #include "qgsvectortilemvtdecoder.h"
 #include "qgsthreadingutils.h"
 #include "qgsproviderregistry.h"
-#include "qgsziputils.h"
-#include "qgsapplication.h"
-#include "qgsmessagelog.h"
-#include "qgsauthmanager.h"
+#include "qgsvectortiledataprovider.h"
+#include "qgsmbtilesvectortiledataprovider.h"
+#include "qgsarcgisvectortileservicedataprovider.h"
+#include "qgsxyzvectortiledataprovider.h"
+#include "qgsvtpkvectortiledataprovider.h"
 
 #include <QUrl>
 #include <QUrlQuery>
@@ -1315,418 +1314,5 @@ void QgsVectorTileLayer::removeSelection()
   mSelectedFeatures.clear();
   emit selectionChanged();
 }
-
-
-//
-// QgsVectorTileDataProvider
-//
-///@cond PRIVATE
-QgsVectorTileDataProvider::QgsVectorTileDataProvider( const QString &uri,
-    const ProviderOptions &options,
-    QgsDataProvider::ReadFlags flags )
-  : QgsDataProvider( uri, options, flags )
-{}
-
-QString QgsVectorTileDataProvider::name() const
-{
-  QGIS_PROTECT_QOBJECT_THREAD_ACCESS
-
-  return QStringLiteral( "vectortile" );
-}
-
-QString QgsVectorTileDataProvider::description() const
-{
-  QGIS_PROTECT_QOBJECT_THREAD_ACCESS
-
-  return QString();
-}
-
-QgsRectangle QgsVectorTileDataProvider::extent() const
-{
-  QGIS_PROTECT_QOBJECT_THREAD_ACCESS
-
-  return QgsRectangle();
-}
-
-bool QgsVectorTileDataProvider::renderInPreview( const PreviewContext &context )
-{
-  QGIS_PROTECT_QOBJECT_THREAD_ACCESS
-
-  // Vector tiles by design are very CPU light to render, so we are much more permissive here compared
-  // with other layer types. (Generally if a vector tile layer has taken more than a few milliseconds to render it's
-  // a result of network requests, and the tile manager class handles these gracefully for us)
-  return context.lastRenderingTimeMs <= 1000;
-}
-
-bool QgsVectorTileDataProvider::supportsAsync() const
-{
-  return false;
-}
-
-QNetworkRequest QgsVectorTileDataProvider::tileRequest( const QgsTileMatrix &, const QgsTileXYZ &, Qgis::RendererUsage ) const
-{
-  return QNetworkRequest();
-}
-
-//
-// QgsXyzVectorTileDataProvider
-//
-
-QgsXyzVectorTileDataProvider::QgsXyzVectorTileDataProvider( const QString &uri, const ProviderOptions &providerOptions, ReadFlags flags )
-  : QgsVectorTileDataProvider( uri, providerOptions, flags )
-{
-  QgsDataSourceUri dsUri;
-  dsUri.setEncodedUri( uri );
-
-  mAuthCfg = dsUri.authConfigId();
-  mHeaders = dsUri.httpHeaders();
-}
-
-QgsVectorTileDataProvider *QgsXyzVectorTileDataProvider::clone() const
-{
-  QGIS_PROTECT_QOBJECT_THREAD_ACCESS
-
-  ProviderOptions options;
-  options.transformContext = transformContext();
-  return new QgsXyzVectorTileDataProvider( dataSourceUri(), options, mReadFlags );
-}
-
-QString QgsXyzVectorTileDataProvider::sourcePath() const
-{
-  QGIS_PROTECT_QOBJECT_THREAD_ACCESS
-
-  QgsDataSourceUri dsUri;
-  dsUri.setEncodedUri( dataSourceUri() );
-  return dsUri.param( QStringLiteral( "url" ) );
-}
-
-bool QgsXyzVectorTileDataProvider::isValid() const
-{
-  QGIS_PROTECT_QOBJECT_THREAD_ACCESS
-
-  return true;
-}
-
-QgsCoordinateReferenceSystem QgsXyzVectorTileDataProvider::crs() const
-{
-  QGIS_PROTECT_QOBJECT_THREAD_ACCESS
-
-  return QgsCoordinateReferenceSystem( QStringLiteral( "EPSG:3857" ) );
-}
-
-bool QgsXyzVectorTileDataProvider::supportsAsync() const
-{
-  return true;
-}
-
-QByteArray QgsXyzVectorTileDataProvider::readTile( const QgsTileMatrix &tileMatrix, const QgsTileXYZ &id, QgsFeedback *feedback ) const
-{
-  return loadFromNetwork( id, tileMatrix, sourcePath(), mAuthCfg, mHeaders, feedback );
-}
-
-QList<QgsVectorTileRawData> QgsXyzVectorTileDataProvider::readTiles( const QgsTileMatrix &tileMatrix, const QVector<QgsTileXYZ> &tiles, QgsFeedback *feedback ) const
-{
-  QList<QgsVectorTileRawData> rawTiles;
-  rawTiles.reserve( tiles.size() );
-  const QString source = sourcePath();
-  for ( QgsTileXYZ id : std::as_const( tiles ) )
-  {
-    if ( feedback && feedback->isCanceled() )
-      break;
-
-    const QByteArray rawData = loadFromNetwork( id, tileMatrix, source, mAuthCfg, mHeaders, feedback );
-    if ( !rawData.isEmpty() )
-    {
-      rawTiles.append( QgsVectorTileRawData( id, rawData ) );
-    }
-  }
-  return rawTiles;
-}
-
-QNetworkRequest QgsXyzVectorTileDataProvider::tileRequest( const QgsTileMatrix &tileMatrix, const QgsTileXYZ &id, Qgis::RendererUsage usage ) const
-{
-  QString urlTemplate = sourcePath();
-
-  if ( urlTemplate.contains( QLatin1String( "{usage}" ) ) )
-  {
-    switch ( usage )
-    {
-      case Qgis::RendererUsage::View:
-        urlTemplate.replace( QLatin1String( "{usage}" ), QLatin1String( "view" ) );
-        break;
-      case Qgis::RendererUsage::Export:
-        urlTemplate.replace( QLatin1String( "{usage}" ), QLatin1String( "export" ) );
-        break;
-      case Qgis::RendererUsage::Unknown:
-        urlTemplate.replace( QLatin1String( "{usage}" ), QString() );
-        break;
-    }
-  }
-
-  const QString url = QgsVectorTileUtils::formatXYZUrlTemplate( urlTemplate, id, tileMatrix );
-
-  QNetworkRequest request( url );
-  QgsSetRequestInitiatorClass( request, QStringLiteral( "QgsXyzVectorTileDataProvider" ) );
-  QgsSetRequestInitiatorId( request, id.toString() );
-
-  request.setAttribute( static_cast<QNetworkRequest::Attribute>( QNetworkRequest::User + 1 ), id.column() );
-  request.setAttribute( static_cast<QNetworkRequest::Attribute>( QNetworkRequest::User + 2 ), id.row() );
-  request.setAttribute( static_cast<QNetworkRequest::Attribute>( QNetworkRequest::User + 3 ), id.zoomLevel() );
-
-  request.setAttribute( QNetworkRequest::CacheLoadControlAttribute, QNetworkRequest::PreferCache );
-  request.setAttribute( QNetworkRequest::CacheSaveControlAttribute, true );
-
-  mHeaders.updateNetworkRequest( request );
-
-  if ( !mAuthCfg.isEmpty() &&  !QgsApplication::authManager()->updateNetworkRequest( request, mAuthCfg ) )
-  {
-    QgsMessageLog::logMessage( tr( "network request update failed for authentication config" ), tr( "Network" ) );
-  }
-
-  return request;
-}
-
-QByteArray QgsXyzVectorTileDataProvider::loadFromNetwork( const QgsTileXYZ &id, const QgsTileMatrix &tileMatrix, const QString &requestUrl, const QString &authid, const QgsHttpHeaders &headers, QgsFeedback *feedback )
-{
-  QString url = QgsVectorTileUtils::formatXYZUrlTemplate( requestUrl, id, tileMatrix );
-  QNetworkRequest nr;
-  nr.setUrl( QUrl( url ) );
-
-  headers.updateNetworkRequest( nr );
-
-  QgsBlockingNetworkRequest req;
-  req.setAuthCfg( authid );
-  QgsDebugMsgLevel( QStringLiteral( "Blocking request: " ) + url, 2 );
-  QgsBlockingNetworkRequest::ErrorCode errCode = req.get( nr, false, feedback );
-  if ( errCode != QgsBlockingNetworkRequest::NoError )
-  {
-    QgsDebugMsg( QStringLiteral( "Request failed: " ) + url );
-    return QByteArray();
-  }
-  QgsNetworkReplyContent reply = req.reply();
-  QgsDebugMsgLevel( QStringLiteral( "Request successful, content size %1" ).arg( reply.content().size() ), 2 );
-  return reply.content();
-}
-
-//
-// QgsMbTilesVectorTileDataProvider
-//
-
-QgsMbTilesVectorTileDataProvider::QgsMbTilesVectorTileDataProvider( const QString &uri, const ProviderOptions &providerOptions, ReadFlags flags )
-  : QgsVectorTileDataProvider( uri, providerOptions, flags )
-{
-
-}
-
-QgsVectorTileDataProvider *QgsMbTilesVectorTileDataProvider::clone() const
-{
-  QGIS_PROTECT_QOBJECT_THREAD_ACCESS
-
-  ProviderOptions options;
-  options.transformContext = transformContext();
-  return new QgsMbTilesVectorTileDataProvider( dataSourceUri(), options, mReadFlags );
-}
-
-QString QgsMbTilesVectorTileDataProvider::sourcePath() const
-{
-  QGIS_PROTECT_QOBJECT_THREAD_ACCESS
-
-  QgsDataSourceUri dsUri;
-  dsUri.setEncodedUri( dataSourceUri() );
-  return dsUri.param( QStringLiteral( "url" ) );
-}
-
-bool QgsMbTilesVectorTileDataProvider::isValid() const
-{
-  QGIS_PROTECT_QOBJECT_THREAD_ACCESS
-
-  return true;
-}
-
-QgsCoordinateReferenceSystem QgsMbTilesVectorTileDataProvider::crs() const
-{
-  QGIS_PROTECT_QOBJECT_THREAD_ACCESS
-
-  return QgsCoordinateReferenceSystem( QStringLiteral( "EPSG:3857" ) );
-}
-
-QByteArray QgsMbTilesVectorTileDataProvider::readTile( const QgsTileMatrix &, const QgsTileXYZ &id, QgsFeedback *feedback ) const
-{
-  QgsDataSourceUri dsUri;
-  dsUri.setEncodedUri( dataSourceUri() );
-
-  QgsMbTiles mbReader( dsUri.param( QStringLiteral( "url" ) ) );
-  mbReader.open();
-  return loadFromMBTiles( mbReader, id, feedback );
-}
-
-QList<QgsVectorTileRawData> QgsMbTilesVectorTileDataProvider::readTiles( const QgsTileMatrix &, const QVector<QgsTileXYZ> &tiles, QgsFeedback *feedback ) const
-{
-  QgsDataSourceUri dsUri;
-  dsUri.setEncodedUri( dataSourceUri() );
-
-  QgsMbTiles mbReader( dsUri.param( QStringLiteral( "url" ) ) );
-  mbReader.open();
-
-  QList<QgsVectorTileRawData> rawTiles;
-  rawTiles.reserve( tiles.size() );
-  for ( QgsTileXYZ id : std::as_const( tiles ) )
-  {
-    if ( feedback && feedback->isCanceled() )
-      break;
-
-    const QByteArray rawData = loadFromMBTiles( mbReader, id, feedback );
-    if ( !rawData.isEmpty() )
-    {
-      rawTiles.append( QgsVectorTileRawData( id, rawData ) );
-    }
-  }
-  return rawTiles;
-}
-
-QByteArray QgsMbTilesVectorTileDataProvider::loadFromMBTiles( QgsMbTiles &mbTileReader, const QgsTileXYZ &id, QgsFeedback *feedback )
-{
-  // MBTiles uses TMS specs with Y starting at the bottom while XYZ uses Y starting at the top
-  const int rowTMS = static_cast<int>( pow( 2, id.zoomLevel() ) - id.row() - 1 );
-  QByteArray gzippedTileData = mbTileReader.tileData( id.zoomLevel(), id.column(), rowTMS );
-  if ( gzippedTileData.isEmpty() )
-  {
-    return QByteArray();
-  }
-
-  if ( feedback && feedback->isCanceled() )
-    return QByteArray();
-
-  QByteArray data;
-  if ( !QgsZipUtils::decodeGzip( gzippedTileData, data ) )
-  {
-    QgsDebugMsg( QStringLiteral( "Failed to decompress tile " ) + id.toString() );
-    return QByteArray();
-  }
-
-  QgsDebugMsgLevel( QStringLiteral( "Tile blob size %1 -> uncompressed size %2" ).arg( gzippedTileData.size() ).arg( data.size() ), 2 );
-  return data;
-}
-
-//
-// QgsVtpkVectorTileDataProvider
-//
-
-QgsVtpkVectorTileDataProvider::QgsVtpkVectorTileDataProvider( const QString &uri, const ProviderOptions &providerOptions, ReadFlags flags )
-  : QgsVectorTileDataProvider( uri, providerOptions, flags )
-{
-
-}
-
-QgsVectorTileDataProvider *QgsVtpkVectorTileDataProvider::clone() const
-{
-  QGIS_PROTECT_QOBJECT_THREAD_ACCESS
-
-  ProviderOptions options;
-  options.transformContext = transformContext();
-  return new QgsVtpkVectorTileDataProvider( dataSourceUri(), options, mReadFlags );
-}
-
-QString QgsVtpkVectorTileDataProvider::sourcePath() const
-{
-  QGIS_PROTECT_QOBJECT_THREAD_ACCESS
-
-  QgsDataSourceUri dsUri;
-  dsUri.setEncodedUri( dataSourceUri() );
-  return dsUri.param( QStringLiteral( "url" ) );
-}
-
-bool QgsVtpkVectorTileDataProvider::isValid() const
-{
-  QGIS_PROTECT_QOBJECT_THREAD_ACCESS
-
-  return true;
-}
-
-QgsCoordinateReferenceSystem QgsVtpkVectorTileDataProvider::crs() const
-{
-  QGIS_PROTECT_QOBJECT_THREAD_ACCESS
-
-  return QgsCoordinateReferenceSystem( QStringLiteral( "EPSG:3857" ) );
-}
-
-QByteArray QgsVtpkVectorTileDataProvider::readTile( const QgsTileMatrix &, const QgsTileXYZ &id, QgsFeedback *feedback ) const
-{
-  QgsDataSourceUri dsUri;
-  dsUri.setEncodedUri( dataSourceUri() );
-  QgsVtpkTiles reader( dsUri.param( QStringLiteral( "url" ) ) );
-  reader.open();
-  return loadFromVtpk( reader, id, feedback );
-}
-
-QList<QgsVectorTileRawData> QgsVtpkVectorTileDataProvider::readTiles( const QgsTileMatrix &, const QVector<QgsTileXYZ> &tiles, QgsFeedback *feedback ) const
-{
-  QgsDataSourceUri dsUri;
-  dsUri.setEncodedUri( dataSourceUri() );
-
-  QgsVtpkTiles reader( dsUri.param( QStringLiteral( "url" ) ) );
-  reader.open();
-
-  QList<QgsVectorTileRawData> rawTiles;
-  rawTiles.reserve( tiles.size() );
-  for ( QgsTileXYZ id : std::as_const( tiles ) )
-  {
-    if ( feedback && feedback->isCanceled() )
-      break;
-
-    const QByteArray rawData = loadFromVtpk( reader, id, feedback );
-    if ( !rawData.isEmpty() )
-    {
-      rawTiles.append( QgsVectorTileRawData( id, rawData ) );
-    }
-  }
-  return rawTiles;
-}
-
-QByteArray QgsVtpkVectorTileDataProvider::loadFromVtpk( QgsVtpkTiles &vtpkTileReader, const QgsTileXYZ &id, QgsFeedback * )
-{
-  const QByteArray tileData = vtpkTileReader.tileData( id.zoomLevel(), id.column(), id.row() );
-  if ( tileData.isEmpty() )
-  {
-    return QByteArray();
-  }
-  return tileData;
-}
-
-//
-// QgsArcGisVectorTileServiceDataProvider
-//
-
-QgsArcGisVectorTileServiceDataProvider::QgsArcGisVectorTileServiceDataProvider( const QString &uri, const QString &sourcePath, const ProviderOptions &providerOptions, ReadFlags flags )
-  : QgsXyzVectorTileDataProvider( uri, providerOptions, flags )
-  , mSourcePath( sourcePath )
-{
-
-}
-
-QgsVectorTileDataProvider *QgsArcGisVectorTileServiceDataProvider::clone() const
-{
-  QGIS_PROTECT_QOBJECT_THREAD_ACCESS
-
-  ProviderOptions options;
-  options.transformContext = transformContext();
-  return new QgsArcGisVectorTileServiceDataProvider( dataSourceUri(), mSourcePath, options, mReadFlags );
-}
-
-QString QgsArcGisVectorTileServiceDataProvider::sourcePath() const
-{
-  QGIS_PROTECT_QOBJECT_THREAD_ACCESS
-
-  return mSourcePath;
-}
-
-bool QgsArcGisVectorTileServiceDataProvider::isValid() const
-{
-  QGIS_PROTECT_QOBJECT_THREAD_ACCESS
-
-  return true;
-}
-
-///@endcond
